/!. WARNING HOTFIX mingw compilation, and BREAK things.
Clues on how to solve this properly are appreciated.
This is in order to get mingw cross-compilation.

Gamma's replacement is either tgamma or gamma2,
upon inspection tgamma seems ok.
Should be safe, but hints to a larger problem perhaps.

Removal of time() in the noise is perhaps catastrophic.
Extreme caution is advised.  We're experimenting.

Bypassing myFirst for the linux version has had no side effect
on the tiny, tiny portion of Essentia we tested.